export: Unify code to export a runbook to Savvy into its own pkg
savvy record and savvy record history both reuse export.ToSavvyArtifact

A future commit will add the ability to export to commands locally to a
markdown file
